Cleanup suivant les indications de quelques inspections de code Intellij Idea
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@58399 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/business/src/main/java/ch/vd/uniregctb/declaration/ordinaire/EnvoiDIsResults.java
@@ -48,7 +48,7 @@ public class EnvoiDIsResults<R extends EnvoiDIsResults> extends JobResults<Long,
 		public final ErreurType raison;
 
 		public Erreur(Long noCtb, Integer officeImpotID, ErreurType raison, String details) {
-			super((noCtb == null ? 0 : noCtb.longValue()), officeImpotID, details);
+			super((noCtb == null ? 0 : noCtb), officeImpotID, details);
 			this.raison = raison;
 		}
